+++ src/main/java/be/tomcools/tombot/model/FacebookMessageDelivery.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 public class FacebookMessageDelivery {
     public List<String> mids;
-    public Integer watermark;
-    public Integer seq;
+    public Long watermark;
+    public Long seq;
 }
