+++ src/main/java/be/tomcools/tombot/model/FacebookMessageRead.java
@@ -4,6 +4,6 @@ package be.tomcools.tombot.model;
  * Created by tomco on 30/10/2016.
  */
 public class FacebookMessageRead {
-    Integer watermark;
-    Integer seq;
+    Long watermark;
+    Long seq;
 }
